make sure rational-checks are done on concrete AlgReals
When we process an AlgReal on the Haskell side, it's not sufficient
that this value is concrete. It should also be something that Haskell
can process, i.e., not a true real number that cannot be processed as a
rational. This commit makes sure proper checks are done to satisfy this
constraint, otherwise treating a given concrete real value as symbolic
as far as constant folding is concerned.