Implement structural equality/ordering over Algebraic Reals
Strictly speaking, there's no way we can implement equality or ordering
on Algebraic Reals in Haskell without going through the entire saga of
representing them precisely. However, our use case only calls for a
structural equality test as we only use these for indexes to Maps,
which has to be an instance of Ord. So, we define our custom versions.

Note that we do *not* use Haskell's Eq and Ord classes for this
purpose, as it would be cheating. Those instances remain faithful to
the semantics, and simply raise an error when they receive true real
numbers.