Allow smt-Lib compilation to take a sat/prove parameter
The compileToSMTLib and generateSMTBenchmarks functions translated the
properties as if they were passed as an argument to 'prove', i.e., by
negating the goal and asking the SMT solver if this negation is
unsatisfiable. This commit allows users to specify whether they want
compilation to be done in the SAT style (i.e., direct), or PROVE style
(i.e., negated), by providing an extra boolean flag.

Addresses issue #56.