Fix select when #elts > indexable
The original implementation of select was buggy in the case where there
were more elements than can be indexed by the indexing element. (For
instance 257 elements with an index of SWord8.) Carefully limit the
number of elements to avoid run-over.

Thanks to Brian Huffman for the report.

Addresses #103; will close after confirmation from Brian.