Improve response when backend solver has no model
In case of “unknown” the backend solver may not have a model to
present. Watch for that case and respond appropriately, instead of
choking.. This is terribly z3 specific I’m afraid, but will have to do
for now; until SMTLib folks come up with a way to query solvers to see
if there’s model available.