Clean-up uninterpreted functions
In particular, do not piggy-back the casts on these; that was a cute
hack to begin with but the code complexity is just not worth it. Also,
it is much easier to represent these separately for translation to
other backends (i.e., C); though for SMTLib they act more like
uninterpreted functions. The current design is a bit more verbose, but
is easier to maintain and understand.