backport to GHC 7.8
This patch unfortunately reverts back to using standalone `instance`
declarations rather than taking advantage of `DeriveAnyClass`, but the
impact isn't too huge.

Also uses the `base-compat` to smooth over much of the warnings from
Prelude changes between 7.8 and 7.10. As a result, you'll see in a few
modules:

```
import Prelude ()
import Prelude.Compat
```

The biggest wart is with overlapping instances for the `Equality` class,
which I smoothed over as much as possible with pragmas at the top of
`Data.SBV`. That's definitely a spot to go in and fix once 7.8 is no
longer supported.