add compatibility shims for 7.8
This tries to be minimally invasive by adding `Compat` modules for the
relevant bits. Of course, unlike a true compat module this doesn't
reimplement the functionality lost in earlier versions, instead I have
it set to emit warnings and return benign values like empty strings and
`0`.

When the next GHC is out, we can drop 7.8 and tear all this stuff back
out :+1: