Simplify constant-table structure
A much better fix for #382. We need not keep track of
negative/positive 0; but rather properly account for their
presence. Also, previously NaNs were never "shared"; surely
this isn't a big deal, but we now share them as well since
there's only one NaN in SMTLib and Haskell. (As opposed to
IEEE754).